+++ include/sdl++/filesystem.hpp
@@ -25,9 +25,9 @@
 
 #include "SDL_filesystem.h"
 
+#include "detail/wrapper.hpp"
 #include "macros.hpp"
 #include "stdinc.hpp"
-#include "utils.hpp"
 
 namespace sdl {
